+++ drivers/tty/serial/amba-pl011.c
@@ -1137,7 +1137,7 @@ static inline void pl011_dma_rx_stop(struct uart_amba_port *uap)
  */
 static void pl011_dma_rx_poll(unsigned long args)
 {
-	struct uart_amba_port *uap = (struct uart_amba_port *)args;
+	struct uart_amba_port *uap = (struct uart_amba_port *)(uintptr_t)args;
 	struct tty_port *port = &uap->port.state->port;
 	struct pl011_dmarx_data *dmarx = &uap->dmarx;
 	struct dma_chan *rxchan = uap->dmarx.chan;
@@ -1259,7 +1259,7 @@ skip_rx:
 		if (uap->dmarx.poll_rate) {
 			init_timer(&(uap->dmarx.timer));
 			uap->dmarx.timer.function = pl011_dma_rx_poll;
-			uap->dmarx.timer.data = (unsigned long)uap;
+			uap->dmarx.timer.data = (uintptr_t)uap;
 			mod_timer(&uap->dmarx.timer,
 				jiffies +
 				msecs_to_jiffies(uap->dmarx.poll_rate));
@@ -1636,7 +1636,7 @@ static unsigned int pl011_tx_empty(struct uart_port *port)
 	    container_of(port, struct uart_amba_port, port);
 	unsigned int status = pl011_read(uap, REG_FR);
 	return status & (uap->vendor->fr_busy | UART01x_FR_TXFF) ?
-							0 : TIOCSER_TEMT;
+							0 : TIOCSER_TEMT;/*[false alarm]:return */
 }
 
 static unsigned int pl011_get_mctrl(struct uart_port *port)
